feat: implement ClientPage and GameOver components, refactor Quiz page for server-side data fetching fix: update Health component to display health as images and modify Modal for game over state fix: enhance Timer component cleanup and adjust option-yellow styling in CSS
