+++ src/app/quiz/[id]/health.tsx
@@ -1,11 +1,23 @@
+import Image from "next/image";
+
 type HealthProps = {
   health: number;
 };
 
-export default function Health(health: HealthProps) {
+export default function Health({ health }: HealthProps) {
   return (
     <main>
-      <div>{health.health}</div>
+      <div style={{ display: "flex" }}>
+        {Array.from({ length: health }, (_, index) => (
+          <Image
+            key={index}
+            src="/images/hearth.png"
+            alt="Heart"
+            width={20}
+            height={20}
+          />
+        ))}
+      </div>
     </main>
   );
 }
