+++ src/app/quiz/[id]/modal.tsx
@@ -1,19 +1,34 @@
 import React from "react";
+import Health from "./health";
 
 interface ModalProps {
   show: boolean;
   message: string;
+  score: string;
+  health: number;
 }
 
-const Modal: React.FC<ModalProps> = ({ show, message }) => {
-  if (!show) {
+const Modal: React.FC<ModalProps> = ({ show, message, score, health }) => {
+  // Pastikan modal tetap muncul jika health < 1
+  if (!show && health >= 1) {
     return null;
   }
 
   return (
     <div style={styles.overlay}>
       <div style={styles.modal}>
-        <p>{message}</p>
+        {health >= 1 ? (
+          <>
+            <Health health={health} />
+            <p>{message}</p>
+          </>
+        ) : (
+          <>
+            <h1>Game Over</h1>
+            <p>your score</p>
+            <h2>{score}</h2>
+          </>
+        )}
       </div>
     </div>
   );
