+++ src/app/quiz/[id]/module.css
@@ -67,6 +67,7 @@ body {
 
 .option-yellow {
   background-color: #ffeb3b;
+  color: black;
 }
 
 .option-green {
