+++ src/app/quiz/[id]/timer.tsx
@@ -8,9 +8,8 @@ const Timer = () => {
       setTimeLeft((prevTime) => prevTime - 1);
     }, 1000);
 
-    // Clean up the interval on component unmount
     return () => clearInterval(timer);
-  }, []); // Empty dependency array ensures this effect runs only once
+  }, []);
 
   return <div>Timer {timeLeft} detik</div>;
 };
